refactor(project): extract the project crate below mcp and daemon
