+++ crates/tracedecay-project/Cargo.toml
@@ -1,0 +1,115 @@
+[package]
+name = "tracedecay-project"
+# Workspace-versioned on purpose: `version::PACKAGE_VERSION` is this crate's
+# `CARGO_PKG_VERSION`, the released version the product runtime validates the
+# generating binary against.
+version.workspace = true
+publish = false
+edition.workspace = true
+license = "MIT"
+description = "Registered TraceDecay project handle: configuration authority, runtime-port wiring, product runtime, and the registered test-runtime fixture"
+repository = "https://github.com/ScriptedAlchemy/tracedecay"
+
+[lib]
+doctest = false
+
+[features]
+# Opt-in profiling. Hotpath macros are no-ops until the *binary* selects the
+# `hotpath/hotpath` backend (cargo unifies the proc-macro's features), so this
+# feature exists only for the `cfg(feature = "hotpath")` sites in this crate.
+hotpath = ["hotpath/hotpath", "hotpath/tokio"]
+
+# Integration fixture surface: `HostAdmissionTestRuntimeV1`, the registered
+# database scaffolding it owns, and the explicit `TraceDecay::*_for_test`
+# constructors that open a project through it. Production code is never gated
+# by this feature; it compiles only under `cfg(test)` or when a consumer asks
+# for it explicitly.
+test-helpers = [
+    "dep:rusqlite",
+    "tracedecay-automation-runtime/test-helpers",
+    "tracedecay-code-index-runtime/test-helpers",
+    "tracedecay-host-admission/test-helpers",
+    "tracedecay-runtime-core/test-helpers",
+    "tracedecay-session-runtime/test-helpers",
+    "tracedecay-sessions/test-helpers",
+    "tracedecay-store-runtime/test-helpers",
+    "tracedecay-global-db/test-helpers",
+    "tracedecay-session-temporal-store/test-helpers",
+    "tracedecay-code-index/test-helpers",
+    "tracedecay-graph-query/test-helpers",
+]
+
+# Mirrors the composition root's `test-transport`: the standalone
+# `TraceDecay::init` / `open` / `open_read_only` entry points route through
+# the registered test runtime instead of the exclusive maintenance lease.
+test-transport = [
+    "test-helpers",
+    "tracedecay-rusqlite-runtime/test-transport",
+    "tracedecay-runtime-core/test-transport",
+    "tracedecay-session-memory/test-transport",
+    "tracedecay-agent-hosts/test-transport",
+    "tracedecay-automation-runtime/test-transport",
+    "tracedecay-dashboard-api/test-transport",
+    "tracedecay-host-admission/test-transport",
+    "tracedecay-application/test-transport",
+    "tracedecay-code-index-runtime/test-transport",
+]
+
+[dependencies]
+hotpath = { workspace = true, features = ["tokio"] }
+# Only the `test-helpers` LCM fixture opens a raw connection.
+rusqlite = { version = "0.40.1", default-features = false, features = ["backup"], optional = true }
+serde_json = "1"
+sha2 = "0.11"
+thiserror = "2"
+tokio = { version = "1", features = ["full"] }
+tracing = "0.1"
+tracedecay-agent-hosts = { path = "../tracedecay-agent-hosts", version = "0.1.0" }
+tracedecay-api = { path = "../tracedecay-api", version = "0.1.0" }
+tracedecay-application = { path = "../tracedecay-application", version = "0.1.0" }
+tracedecay-automation-runtime = { path = "../tracedecay-automation-runtime", version = "0.1.0" }
+tracedecay-code-index = { path = "../tracedecay-code-index", version = "0.1.0", default-features = false }
+tracedecay-code-index-runtime = { path = "../tracedecay-code-index-runtime", version = "0.1.0", default-features = false }
+tracedecay-configuration = { path = "../tracedecay-configuration", version = "0.1.0" }
+tracedecay-contracts = { path = "../tracedecay-contracts", version = "0.1.0" }
+tracedecay-daemon-identity = { path = "../tracedecay-daemon-identity", version = "0.1.0" }
+tracedecay-daemon-protocol = { path = "../tracedecay-daemon-protocol", version = "0.1.0" }
+tracedecay-dashboard-api = { path = "../tracedecay-dashboard-api", version = "0.1.0" }
+tracedecay-domain = { path = "../tracedecay-domain", version = "0.1.0" }
+tracedecay-global-db = { path = "../tracedecay-global-db", version = "0.1.0" }
+tracedecay-graph-db = { path = "../tracedecay-graph-db", version = "0.1.0" }
+tracedecay-graph-query = { path = "../tracedecay-graph-query", version = "0.1.0" }
+tracedecay-hooks = { path = "../tracedecay-hooks", version = "0.1.0" }
+tracedecay-host-admission = { path = "../tracedecay-host-admission", version = "0.1.0" }
+tracedecay-lcm = { path = "../tracedecay-lcm", version = "0.1.0" }
+tracedecay-runtime-core = { path = "../tracedecay-runtime-core", version = "0.1.0" }
+tracedecay-rusqlite-runtime = { path = "../tracedecay-rusqlite-runtime", version = "0.1.0" }
+tracedecay-session-memory = { path = "../tracedecay-session-memory", version = "0.1.0" }
+tracedecay-session-runtime = { path = "../tracedecay-session-runtime", version = "0.1.0" }
+tracedecay-session-temporal-store = { path = "../tracedecay-session-temporal-store", version = "0.1.0" }
+tracedecay-sessions = { path = "../tracedecay-sessions", version = "0.1.0" }
+tracedecay-source-edit = { path = "../tracedecay-source-edit", version = "0.1.0" }
+tracedecay-store = { path = "../tracedecay-store", version = "0.1.0" }
+tracedecay-store-runtime = { path = "../tracedecay-store-runtime", version = "0.1.0" }
+tracedecay-temporal-query = { path = "../tracedecay-temporal-query", version = "0.1.0" }
+tracedecay-tool-catalog = { path = "../tracedecay-tool-catalog", version = "0.1.0" }
+
+[dev-dependencies]
+rusqlite = { version = "0.40.1", default-features = false, features = ["backup"] }
+tempfile = "3"
+tokio = { version = "1", features = ["full", "test-util"] }
+tracedecay-semantic = { path = "../tracedecay-semantic", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-semantic-contracts.workspace = true
+tracedecay-application = { path = "../tracedecay-application", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-automation-runtime = { path = "../tracedecay-automation-runtime", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-code-index = { path = "../tracedecay-code-index", version = "0.1.0", default-features = false, features = ["test-helpers"] }
+tracedecay-code-index-runtime = { path = "../tracedecay-code-index-runtime", version = "0.1.0", default-features = false, features = ["test-helpers", "test-transport"] }
+tracedecay-dashboard-api = { path = "../tracedecay-dashboard-api", version = "0.1.0", features = ["test-transport"] }
+tracedecay-global-db = { path = "../tracedecay-global-db", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-graph-query = { path = "../tracedecay-graph-query", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-host-admission = { path = "../tracedecay-host-admission", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-runtime-core = { path = "../tracedecay-runtime-core", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-session-runtime = { path = "../tracedecay-session-runtime", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-sessions = { path = "../tracedecay-sessions", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-store-runtime = { path = "../tracedecay-store-runtime", version = "0.1.0", features = ["test-helpers"] }
+tracedecay-temporal-query = { path = "../tracedecay-temporal-query", version = "0.1.0", features = ["test-helpers"] }
