+++ crates/tracedecay-project/src/project/automation_context.rs
