+++ crates/tracedecay-project/src/test_support/mod.rs
@@ -1,0 +1,4 @@
+//! Test-only fixtures composed over the project handle. Compiled under
+//! `cfg(test)` or `test-helpers`, never into a default library build.
+
+pub mod host_admission;
