+++ crates/tracedecay/Cargo.toml
@@ -106,6 +106,7 @@ hotpath = [
     "tracedecay-hooks/hotpath",
     "tracedecay-lcm/hotpath",
     "tracedecay-mcp/hotpath",
+    "tracedecay-project/hotpath",
     "tracedecay-query/hotpath",
     "tracedecay-rusqlite-runtime/hotpath",
     "tracedecay-sdk/hotpath",
@@ -205,6 +206,7 @@ lang-lean = ["tracedecay-code-index/lang-lean"]
 # out of a default or `production` build: they compile only under `cfg(test)`
 # or when a consumer asks for them explicitly.
 test-helpers = [
+    "tracedecay-project/test-helpers",
     "tracedecay-automation-runtime/test-helpers",
     "tracedecay-code-index-runtime/test-helpers",
     "tracedecay-daemon-service/test-helpers",
@@ -218,6 +220,7 @@ test-helpers = [
 
 test-transport = [
     "test-helpers",
+    "tracedecay-project/test-transport",
     "tracedecay-rusqlite-runtime/test-transport",
     "tracedecay-runtime-core/test-transport",
     "tracedecay-session-memory/test-transport",
@@ -338,6 +341,8 @@ tracedecay-graph-query = { path = "../tracedecay-graph-query", version = "0.1.0"
 tracedecay-policy = { path = "../tracedecay-policy", version = "0.1.0" }
 tracedecay-privacy = { path = "../tracedecay-privacy", version = "0.1.0" }
 tracedecay-private-fs = { path = "../tracedecay-private-fs", version = "0.1.0" }
+# Workspace-versioned like this crate: it owns `version::PACKAGE_VERSION`.
+tracedecay-project = { path = "../tracedecay-project", version = "0.1.0-beta.37" }
 tracedecay-query = { path = "../tracedecay-query", version = "0.1.0" }
 tracedecay-runtime-core = { path = "../tracedecay-runtime-core", version = "0.1.0" }
 tracedecay-search-eval = { path = "../tracedecay-search-eval", version = "0.1.0" }
@@ -427,6 +432,12 @@ tracedecay-dashboard-api = { path = "../tracedecay-dashboard-api", version = "0.
 # behind tracedecay-mcp's test-transport feature, so root unit tests must
 # enable that feature without putting it on the production CLI lane.
 tracedecay-mcp = { path = "../tracedecay-mcp", version = "0.1.0", features = ["test-transport"] }
+# The registered host-admission test runtime and the explicit
+# `TraceDecay::*_for_test` constructors root unit tests open projects through.
+# Only `test-helpers`: the project crate's `test-transport` reroutes the plain
+# `TraceDecay::init`/`open` API onto the fixture, and the root's
+# `test-helpers`-only suites exercise the production standalone path.
+tracedecay-project = { path = "../tracedecay-project", version = "0.1.0-beta.37", features = ["test-helpers"] }
 tracedecay-runtime-core = { path = "../tracedecay-runtime-core", version = "0.1.0", features = ["test-helpers"] }
 tracedecay-sessions = { path = "../tracedecay-sessions", version = "0.1.0", features = ["test-helpers"] }
 tracedecay-host-admission = { path = "../tracedecay-host-admission", version = "0.1.0", features = ["test-helpers"] }
