+++ crates/tracedecay/tests/mcp_suite/mcp_handler_test/status_runtime_test.rs
@@ -116,7 +116,7 @@ async fn status_reports_daemon_owned_partial_history_catch_up() {
         None,
         None,
         tracedecay::mcp::tools::ToolCallRegistryOptions::with_session_authorities(
-            runtime.mcp_session_authorities(),
+            tracedecay::test_support::host_admission::mcp_session_authorities(&runtime),
         )
         .admit_opened_project(&cg)
         .unwrap(),
@@ -180,7 +180,7 @@ async fn runtime_exposes_cursor_ingest_health_for_daemon_owned_doctor_checks() {
         None,
         None,
         tracedecay::mcp::tools::ToolCallRegistryOptions::with_session_authorities(
-            runtime.mcp_session_authorities(),
+            tracedecay::test_support::host_admission::mcp_session_authorities(&runtime),
         )
         .admit_opened_project(&cg)
         .unwrap(),
